refactor(plugin): extract feishu markdown utils
