+++ docs/superpowers/specs/2026-08-10-plugin-feishu-markdown-utils-design.md
@@ -1,0 +1,43 @@
+# 飞书 Markdown 纯格式化模块拆分设计
+
+## 背景
+
+插件 `src/main.js` 同时承载飞书网络接入和约 400 行纯文本清洗规则。后者只接收 Markdown、标题或行数组，输出字符串或布尔值，不读取设置、网络、文件、账户和真实文档，适合单独成为可测试模块。
+
+## 方案比较
+
+### 方案 A：原样迁移纯格式化函数（采用）
+
+把现有规范化、噪声过滤、目录清理、表格修复、代码块整理、尾部推荐裁剪和截断判断原样迁入 `src/feishu-markdown-utils.js`。主文件只导入四个实际跨边界调用的函数。优点是行为面最小、可用固定样本证明等价。
+
+### 方案 B：迁移时顺便优化历史规则
+
+可能减少硬编码，但会把架构重构和用户可见行为变化混在一起，无法判断回归来自迁移还是规则变化，因此拒绝。
+
+### 方案 C：只迁移最终入口，内部 helper 留在主文件
+
+会形成反向依赖或重复实现，不能建立独立模块边界，因此拒绝。
+
+## 模块边界
+
+新模块公开：
+
+- `normalizeFeishuMarkdownLine`
+- `formatFeishuHeadingLine`
+- `postProcessFeishuMarkdown`
+- `isFeishuMarkdownLikelyTruncated`
+
+其余 helper 保持模块内部。网络请求、OAuth、网页渲染、官方 API、图片处理、笔记写入和刷新编排仍由 `src/main.js` 负责。
+
+## 等价性证明
+
+1. 零宽字符和列表行规范化输出保持不变。
+2. 中文章节标题、普通文本和命令行的标题格式保持不变。
+3. 飞书壳层噪声、目录块、评论尾部、表格残片和推荐尾部处理保持不变。
+4. 语言行与命令行仍组合成相同 fenced code block。
+5. 完整正文和疑似截断正文的布尔判断保持不变。
+6. 现有调用顺序、参数和正式单文件 bundle 行为保持不变。
+
+## 风险与回退
+
+主要风险是漏迁内部 helper、改变正则字面量或遗漏主文件调用点。任何固定样本、主回归、构建漂移或重复构建哈希异常即停止并丢弃本阶段 Worktree；第八阶段提交保持可直接回退。
